% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{ApplyDictToData}
\alias{ApplyDictToData}
\title{ApplyDictToData}
\usage{
ApplyDictToData(dict, data)
}
\arguments{
\item{dict}{data frame with dictionary information}

\item{data}{data frame with data}
}
\value{
data frame with data
}
\description{
apply dictionary to data
}
