% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_explain.R
\name{autoplot.t2e_explain}
\alias{autoplot.t2e_explain}
\title{Autoplot method for t2e_explain objects}
\usage{
\method{autoplot}{t2e_explain}(object, what = c("importance"), ...)
}
\arguments{
\item{object}{A `t2e_explain` object.}

\item{what}{What to plot (currently only "importance" is supported).}

\item{...}{Additional arguments passed to ggplot2.}
}
\value{
A ggplot2 object.
}
\description{
Autoplot method for t2e_explain objects
}
