% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaderboard.R
\name{autoplot.t2e_leaderboard}
\alias{autoplot.t2e_leaderboard}
\title{Plot model leaderboard}
\usage{
\method{autoplot}{t2e_leaderboard}(object, metric = NULL, ...)
}
\arguments{
\item{object}{A \code{t2e_leaderboard} object.}

\item{metric}{The metric to plot. Defaults to the first metric available.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot model leaderboard
}
