% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_steps.R
\name{bake_data_recipe}
\alias{bake_data_recipe}
\title{bake_data_recipe}
\usage{
bake_data_recipe(model_recipe, data)
}
\arguments{
\item{model_recipe}{initiated model recipe}

\item{data}{data to be baked}
}
\value{
baked data
}
\description{
bake data using a recipe
}
