% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_shallownn.R
\name{baseline_cumhaz_stepfun}
\alias{baseline_cumhaz_stepfun}
\title{(Helper) Baseline Hazard Step Function}
\usage{
baseline_cumhaz_stepfun(baseline_df)
}
\arguments{
\item{baseline_df}{Data frame with columns `time` and `cumhaz`.}
}
\value{
A step function suitable for evaluating cumulative hazard at arbitrary times.
}
\description{
Generates a right-continuous step function for the baseline cumulative hazard.
}
\keyword{internal}
