% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{bindDicts}
\alias{bindDicts}
\title{bindDict}
\usage{
bindDicts(Dict1, Dict2)
}
\arguments{
\item{dict1}{data frame with dictionary information}

\item{dict2}{data frame with dictionary information}
}
\value{
data frame with dictionary information
}
\description{
Utility function to combine two dictionaries in a serial fashion
}
