% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_target_builder.R
\name{buildObservedCIFMatrix}
\alias{buildObservedCIFMatrix}
\title{Build Observed CIF Matrix}
\usage{
buildObservedCIFMatrix(data, timevar, eventvar, cause_of_interest, eval_times)
}
\arguments{
\item{data}{Training data containing time and event variables}

\item{timevar}{Name of time variable}

\item{eventvar}{Name of event variable (0, 1, 2, ...)}

\item{cause_of_interest}{Cause of interest (1, 2, etc.)}

\item{eval_times}{Times at which to evaluate CIF probabilities}
}
\value{
Matrix of observed CIF probabilities (rows=times, cols=observations)
}
\description{
Construct observed cumulative incidence matrix from competing risks data
}
