% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_target_builder.R
\name{buildObservedSurvivalMatrix}
\alias{buildObservedSurvivalMatrix}
\title{Build Observed Survival Matrix}
\usage{
buildObservedSurvivalMatrix(data, timevar, eventvar, eval_times)
}
\arguments{
\item{data}{Training data containing time and event variables}

\item{timevar}{Name of time variable}

\item{eventvar}{Name of event variable (0/1)}

\item{eval_times}{Times at which to evaluate survival probabilities}
}
\value{
Matrix of observed survival probabilities (rows=times, cols=observations)
}
\description{
Construct observed survival probability matrix from training data
}
