% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HlaDataFunctions.R
\name{Calc6MatchingDists}
\alias{Calc6MatchingDists}
\title{Calc6MatchingDists}
\usage{
Calc6MatchingDists(
  d_typ1,
  d_typ2,
  r_typ1,
  r_typ2,
  field = "A",
  allelenames,
  DistAlleles
)
}
\description{
Gets 6 sequence based distance matrices for a given set of donor
recipient alleles
}
