% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HlaDataFunctions.R
\name{CalcHLAGenotypeFeatures}
\alias{CalcHLAGenotypeFeatures}
\title{CalcHLAGenotypeFeatures}
\usage{
CalcHLAGenotypeFeatures(typ1, typ2, field = "A", allelenames, FeatAlleles)
}
\description{
Given sequence based features  calculate features for a genotype
at a HLA field
}
