% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecTools.R
\name{CalculateExpectedTimeLost}
\alias{CalculateExpectedTimeLost}
\title{CalculateExpectedTimeLost}
\usage{
CalculateExpectedTimeLost(PredictedCurves, modeltypes, times, UL, LL)
}
\description{
Predicted time lost frm survival or cr curves
}
