% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{CatVarsDict}
\alias{CatVarsDict}
\title{CatVarsDict}
\usage{
CatVarsDict(dict)
}
\arguments{
\item{dict}{data frame with dictionary information (must have 'Type' column).}
}
\value{
data frame containing only rows where Type is "Categorical".
}
\description{
Utility function to subset dictionary for categorical variables.
}
