% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{CharVarsDict}
\alias{CharVarsDict}
\title{CharVarsDict}
\usage{
CharVarsDict(dict)
}
\arguments{
\item{dict}{data frame with dictionary information}
}
\value{
data frame with dictionary information
}
\description{
Utility function to get only the Character variables
}
