% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{ConVarsDict}
\alias{ConVarsDict}
\title{ConVarsDict}
\usage{
ConVarsDict(dict)
}
\arguments{
\item{dict}{data frame with dictionary information (must have 'Type' column).}
}
\value{
data frame containing only rows where Type is "Continuous".
}
\description{
Utility function to subset dictionary for continuous variables.
}
