% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_persistence.R
\name{CREnsemble}
\alias{CREnsemble}
\title{Create CREnsemble S3 Class}
\usage{
CREnsemble(ensemble_list)
}
\arguments{
\item{ensemble_list}{List returned by RunCRModels}
}
\value{
Object of class "CREnsemble"
}
\description{
Converts ensemble model output to S3 class for better organization
}
