% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_bart.R
\name{CRModel_BART}
\alias{CRModel_BART}
\title{CRModel_BART}
\usage{
CRModel_BART(
  data,
  expvars,
  timevar,
  eventvar,
  event_codes = NULL,
  K = 10,
  ntree = 100,
  ndpost = 1000,
  nskip = 100,
  keepevery = 10,
  verbose = FALSE
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (coded 0,1,2 where 0=censored, 1=event of interest, 2=competing event)}

\item{event_codes}{character or numeric vector identifying the event code(s) to
model. BART competing risks currently supports a single event code. If NULL
(default), the first non-zero event code observed in the data is used.}

\item{K}{parameter 'K' for BART (number of time points for discrete approximation, default: 10)}

\item{ntree}{number of trees in BART model (default: 100)}

\item{ndpost}{number of posterior draws to save (default: 1000)}

\item{nskip}{number of MCMC iterations to discard as burn-in (default: 100)}

\item{keepevery}{interval at which to keep posterior draws (default: 10)}

\item{verbose}{logical, print progress messages (default: FALSE)}
}
\value{
a list with the following components:
  \item{bart_model}{the fitted BART model object from BART::crisk.bart}
  \item{times}{vector of unique event times in the training data}
  \item{varprof}{variable profile list containing factor levels and numeric ranges}
  \item{model_type}{character string "cr_bart"}
  \item{expvars}{character vector of explanatory variables used}
  \item{timevar}{character name of time variable}
  \item{eventvar}{character name of event variable}
  \item{event_codes}{character vector of event codes included in the model}
  \item{event_codes_numeric}{numeric vector of event codes included}
  \item{default_event_code}{character scalar for the default event code}
  \item{time_range}{vector with min and max observed event times}
  \item{x_train}{design matrix for the training data}
  \item{times_train}{time variable values from training data}
  \item{delta_train}{event variable values from training data}
}
\description{
Fit a BART model for competing risks outcomes using BART::crisk.bart.
}
