% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_gam.R
\name{CRModel_GAM}
\alias{CRModel_GAM}
\title{CRModel_GAM}
\usage{
CRModel_GAM(
  data,
  expvars,
  timevar,
  eventvar,
  failcode = 1,
  shrinkTreshold = 10,
  ntimes = 50,
  verbose = FALSE
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (coded 0=censored, 1=cause1, 2=cause2, etc.)}

\item{failcode}{integer, the code for the event of interest (default: 1)}

\item{shrinkTreshold}{integer value, minimum number of factor levels for factor variables to be considered for shrinkage ('re' basis).}

\item{ntimes}{integer, number of time points to use for prediction grid (default: 50)}

\item{verbose}{logical, print progress messages (default: FALSE)}
}
\value{
a list with the following components:
  \item{gam_model}{the fitted cause-specific GAM model object from mgcv::gam}
  \item{times}{vector of unique event times in the training data for the event of interest}
  \item{varprof}{variable profile list containing factor levels and numeric ranges}
  \item{model_type}{character string "cr_gam"}
  \item{expvars}{character vector of explanatory variables used}
  \item{timevar}{character name of time variable}
  \item{eventvar}{character name of event variable}
  \item{failcode}{the event code for the outcome of interest}
  \item{time_range}{vector with min and max observed event times}
}
\description{
Fit a GAM model for competing risks outcomes using cause-specific modeling.
}
