% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_rulefit.R
\name{CRModel_rulefit}
\alias{CRModel_rulefit}
\title{CRModel_rulefit}
\usage{
CRModel_rulefit(
  data,
  expvars,
  timevar,
  eventvar,
  failcode = 1,
  ntree = 300,
  nsample = 300,
  keepvars = NULL,
  cuttimes = NULL,
  alpha = 0.5,
  maxit = 2000,
  ...
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (coded 0,1,2 where 0=censored, 1=event of interest, 2=competing event)}

\item{failcode}{integer, the code for the event of interest (default: 1)}

\item{ntree}{number of trees to fit to extract rules (default: 300)}

\item{nsample}{number of samples for each tree (default: 300)}

\item{keepvars}{these variables will be used in each bagging iteration}

\item{cuttimes}{these cut times are used for calculating pseudo observation}

\item{alpha}{numeric, the elastic net mixing parameter for glmnet (default: 0.5)}

\item{maxit}{integer, maximum number of iterations for glmnet (default: 2000)}

\item{...}{additional parameters passed to glmnet::cv.glmnet}
}
\value{
a list with the following components:
  \item{rulefit_model}{the fitted RuleFit model object}
  \item{times}{vector of unique event times in the training data}
  \item{varprof}{variable profile list containing factor levels and numeric ranges}
  \item{model_type}{character string "cr_rulefit"}
  \item{expvars}{character vector of explanatory variables used}
  \item{timevar}{character name of time variable}
  \item{eventvar}{character name of event variable}
  \item{failcode}{the event code for the outcome of interest}
  \item{time_range}{vector with min and max observed event times}
  \item{ctreelist}{list of fitted tree models}
  \item{ruleslist}{list of extracted rules from trees}
}
\description{
Fit a rulefit model for competing risks outcomes using RuleFit algorithm
}
