% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipesDataSteps.R
\name{data_recipe_corstep}
\alias{data_recipe_corstep}
\title{data_recipe_corstep}
\usage{
data_recipe_corstep(model_recipe, treshold = 0.9)
}
\arguments{
\item{model_recipe}{initiated model recipe}

\item{treshold}{removal treshold}
}
\value{
return a  recipe object
}
\description{
Remove variables based on correlation
}
