% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{DateVarstoDateDict}
\alias{DateVarstoDateDict}
\title{DateVarstoDateDict}
\usage{
DateVarstoDateDict(dict, datevars = NULL)
}
\arguments{
\item{dict}{data frame with dictionary information}

\item{datevars}{character vector of variable names}
}
\value{
data frame with dictionary information
}
\description{
Utility function to convert date vars to Date
}
