% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{DetectIdVarsDict}
\alias{DetectIdVarsDict}
\title{DetectIdVarsDict}
\usage{
DetectIdVarsDict(dict, idvars = NULL)
}
\arguments{
\item{dict}{data frame with dictionary information}
}
\value{
character vector of variable names
}
\description{
Utility function to get only the id variables
}
