% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_detection.R
\name{DetectNAStringsDict}
\alias{DetectNAStringsDict}
\title{DetectNAStringsDict}
\usage{
DetectNAStringsDict(dict, knownNaLabels = NULL, knownNaValues = NULL)
}
\arguments{
\item{dict}{data frame with dictionary information (must have 'Label' and 'Value' columns).}

\item{knownNaLabels}{Optional character vector of additional strings in 'Label' column to consider NA.}

\item{knownNaValues}{Optional character vector of additional strings in 'Value' column to consider NA.}
}
\value{
data frame with added 'NACats' column ("NAVAL" or "VALUE").
}
\description{
Utility function to detect potential NA representations in dictionary 'Label' and 'Value' columns.
Adds an 'NACats' column to the dictionary, marking rows as "NAVAL" if suspected NA, otherwise "VALUE".
}
