% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\docType{data}
\name{DICT}
\alias{DICT}
\title{A prepared dictionary for HSCT data from CIBMTR}
\format{
An object of class \code{data.frame} with 2716 rows and 8 columns.
}
\usage{
DICT
}
\description{
A prepared dictionary for HSCT data from CIBMTR
}
\keyword{datasets}
