% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{droplevelsoffactorsData}
\alias{droplevelsoffactorsData}
\title{droplevelsoffactorsData}
\usage{
droplevelsoffactorsData(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame with unused factor levels dropped.
}
\description{
Drop unused factor levels from all factor columns in a data frame.
}
