% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_rulefit.R
\name{extract_rules_from_party}
\alias{extract_rules_from_party}
\title{(Helper) Extract Rules from a party Object}
\usage{
extract_rules_from_party(x)
}
\arguments{
\item{x}{A party object.}
}
\value{
A named character vector where names are the node IDs and values are the rules.
}
\description{
This helper function traverses a party tree
and extracts the decision rules for each terminal node.
}
\keyword{internal}
