% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{filllasttoDict}
\alias{filllasttoDict}
\title{filllasttoDict}
\usage{
filllasttoDict(dict)
}
\arguments{
\item{dict}{data frame with dictionary information}
}
\value{
data frame with dictionary information
}
\description{
Utility function to put dictionary in a special format, fill forward
}
