% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{finddatevars}
\alias{finddatevars}
\title{finddatevars}
\usage{
finddatevars(names, knowndatevars = NULL)
}
\arguments{
\item{names}{character vector of variable names}

\item{knowndatevars}{character vector of known date variables}
}
\value{
character vector of variable names
}
\description{
Utility function to find possible date variables
}
