% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{findvarsnamesthatrepeatData}
\alias{findvarsnamesthatrepeatData}
\title{findvarsnamesthatrepeatData}
\usage{
findvarsnamesthatrepeatData(data)
}
\arguments{
\item{data}{data frame with data}
}
\value{
a vector with variable names that are similar to other variable names


 @export
}
\description{
Find variable names that are similar to other variable names
}
