% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecTools.R
\name{GetAllConditionalCategoriesData}
\alias{GetAllConditionalCategoriesData}
\title{GetAllCondidtionalCategoriesData}
\usage{
GetAllConditionalCategoriesData(
  data,
  conditions = NULL,
  minmatch = 8,
  mustmatchcols = NULL,
  freevars = NULL
)
}
\description{
Get possible conditional categories  from data
all conditions are equality conditions expressed as data.frame
}
