% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HlaDataFunctions.R
\name{GetAlleleFeatures}
\alias{GetAlleleFeatures}
\title{GetAlleleFeatures}
\usage{
GetAlleleFeatures(alleles, field = "A", allelenames)
}
\description{
Creates a list that includes an allelesindata list for alleles as they match
to allelenames and a design matrix for these matched alleles
}
