% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{getcharcolsData}
\alias{getcharcolsData}
\title{getcharcolsData}
\usage{
getcharcolsData(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Character vector with names of character columns.
}
\description{
Utility function to get the names of character columns in a data frame.
}
