% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_shallownn.R
\name{hazard_increment_matrix}
\alias{hazard_increment_matrix}
\title{(Helper) Hazard Increment Matrix}
\usage{
hazard_increment_matrix(step_fun, times, risk_scores)
}
\arguments{
\item{step_fun}{Step function of baseline cumulative hazard.}

\item{times}{Vector of evaluation times (including time 0).}

\item{risk_scores}{Numeric vector of risk scores exp(eta(x)).}
}
\value{
A matrix with rows = intervals between times and columns = observations.
}
\description{
Creates a matrix of cumulative hazard increments for each observation over a common time grid.
}
\keyword{internal}
