% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{ImputeMissinRecordsData}
\alias{ImputeMissinRecordsData}
\title{ImputeMissinRecordsData}
\usage{
ImputeMissinRecordsData(data, dontuse = NULL, ...)
}
\arguments{
\item{data}{data set to be imputed}

\item{dontuse}{the variables not to use when imputing,
thesev ariables will not be used for imputation or for imputing other
 variables.}

\item{...}{Arguments passed to \code{ranger::ranger()}. If the data set is large,
better use less trees (e.g. \code{num.trees = 20}) and/or a low value of
\code{sample.fraction}.
The following arguments are incompatible, amongst others:
\code{write.forest}, \code{probability}, \code{split.select.weights},
\code{dependent.variable.name}, and \code{classification}.}
}
\value{
Imputed data
}
\description{
Imputation of data using missRanger
}
\examples{
\dontrun{
DATAImputed<-DATA\%>\%ImputeMissinRecordsData()
}

}
\seealso{
[missRanger::missRanger()] which this function wraps.
}
