% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_persistence.R
\name{is.CREnsemble}
\alias{is.CREnsemble}
\title{Check if Object is CREnsemble}
\usage{
is.CREnsemble(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if Object is CREnsemble
}
