% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_persistence.R
\name{is.SurvEnsemble}
\alias{is.SurvEnsemble}
\title{Check if Object is SurvEnsemble}
\usage{
is.SurvEnsemble(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if Object is SurvEnsemble
}
