% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{loaddeepsurv}
\alias{loaddeepsurv}
\title{loaddeepsurv}
\usage{
loaddeepsurv(filename)
}
\arguments{
\item{filename}{the name of the file to load the model from}
}
\value{
a SurvModel_deepsurv model
}
\description{
Load a SurvModel_deepsurv model using load_model_tf
}
