% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{MakelabelledSASDict}
\alias{MakelabelledSASDict}
\title{MakelabelledSASDict}
\usage{
MakelabelledSASDict(fileloc)
}
\arguments{
\item{fileloc}{location of the sas file}
}
\value{
a labelled dictionary


 @export
}
\description{
Make a dictionary from labelled data
}
