% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipesDataSteps.R
\name{minimal_data_recipe}
\alias{minimal_data_recipe}
\title{minimal_data_recipe}
\usage{
minimal_data_recipe(
  model_recipe,
  pmiss = 0.3,
  pother = 0.05,
  dummy = TRUE,
  onehot = FALSE
)
}
\arguments{
\item{model_recipe}{initiated model recipe}

\item{pmiss}{percentage missing to delete a variable}

\item{pother}{percentage treshold to combine fator levels as other}

\item{dummy}{use dummy coding or not}

\item{onehot}{whether onehot encoding is used or not (will miss one category)}
}
\value{
return a  recipe object
}
\description{
Minimal data recipe
}
