% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{ml4t2e_calibrate}
\alias{ml4t2e_calibrate}
\title{Calibrate a fitted model for conformal prediction}
\usage{
ml4t2e_calibrate(object, data)
}
\arguments{
\item{object}{A \code{t2e_fit} object.}

\item{data}{A data frame to be used for calibration. Must contain the same time/event columns as the training data.}

\item{split}{Optional numeric (0 < split < 1). If provided, and \code{data} is NULL, the original training data (stored in \code{object$task$data})
will be split, with \code{split} fraction used for calibration. \emph{Note: this requires re-fitting if the model was already trained on all data,
which this function does NOT do. This argument is reserved for future pipelines where fitting and calibration are coupled.}
Currently, you must provide \code{data}.}
}
\value{
A \code{t2e_fit} object with a \verb{$conformal} slot containing scores and weights.
}
\description{
Computes conformal scores on a calibration set and attaches them to the fit object.
}
