% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{ml4t2e_fit_censoring}
\alias{ml4t2e_fit_censoring}
\title{Fit Censoring Model for Conformal Prediction}
\usage{
ml4t2e_fit_censoring(data, time_col, event_col)
}
\arguments{
\item{data}{A data frame.}

\item{time_col}{Character, name of time column.}

\item{event_col}{Character, name of event column.}
}
\value{
A `survfit` object.
}
\description{
Fits a Kaplan-Meier estimator for the censoring distribution.
}
\keyword{internal}
