% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_wrappers.R
\name{ml4t2e_load_pipeline}
\alias{ml4t2e_load_pipeline}
\title{Load an ml4time2event pipeline}
\usage{
ml4t2e_load_pipeline(file)
}
\arguments{
\item{file}{Path to an `.rds` file created by [ml4t2e_save_pipeline()].}
}
\value{
A reconstructed `ml4t2e_pipeline` object.
}
\description{
Load an ml4time2event pipeline
}
