% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{NumIdVarstoCharDict}
\alias{NumIdVarstoCharDict}
\title{NumIdVarstoCharDict}
\usage{
NumIdVarstoCharDict(dict, idvars = c("`#`", "pseudoid", "pseudoccn"))
}
\arguments{
\item{dict}{data frame with dictionary information}

\item{idvars}{character vector of variable names}
}
\value{
data frame with dictionary information
}
\description{
Utility function to convert id vars to Character
}
