% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{NumVarstCatsData}
\alias{NumVarstCatsData}
\title{NumVarstCatsData}
\usage{
NumVarstCatsData(data, numgroups = NULL, cuts = NULL, min_unique_vals = 5)
}
\arguments{
\item{data}{Data frame containing numeric variables to categorize.}

\item{numgroups}{Integer. If specified, numeric variables (with enough unique values) are cut into this many quantile groups.}

\item{cuts}{Numeric vector. If specified, numeric variables are cut using these specific values as breaks. `numgroups` is ignored if `cuts` is provided.}

\item{min_unique_vals}{Integer. Minimum number of unique non-NA values a numeric variable must have to be categorized (default: 5).}
}
\value{
Data frame with specified numeric variables converted to factors.
}
\description{
Convert numeric variables to categorical (factor) variables based on quantiles or specified cuts.
}
\details{
NumVarstCatsData
}
