% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{OneAgainstRestCorDataSummmary}
\alias{OneAgainstRestCorDataSummmary}
\title{OneAgainstRestCorDataSummmary}
\usage{
OneAgainstRestCorDataSummmary(data)
}
\arguments{
\item{data}{data frame with data}
}
\value{
a vector with correlations of each variable with the rest

 @export
}
\description{
Find variables correlated most with the rest
}
