% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{OneValCat2NumDict}
\alias{OneValCat2NumDict}
\title{OneValCat2NumDict}
\usage{
OneValCat2NumDict(dict, vars = NULL)
}
\arguments{
\item{dict}{data frame with dictionary information.}

\item{vars}{character vector of variable names to potentially correct. If NULL, checks all variables.}
}
\value{
data frame with corrected 'Type' information.
}
\description{
Utility function to correct the 'Type' for variables that appear only once
in the dictionary (implying continuous) but are marked as 'Categorical'.
This assumes single-row entries correspond to continuous variables.
}
