% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{pairwiserelationshipsDataSummmary}
\alias{pairwiserelationshipsDataSummmary}
\title{pairwiserelationshipsDataSummmary}
\usage{
pairwiserelationshipsDataSummmary(data)
}
\arguments{
\item{data}{data frame with data}
}
\value{
a matrix with variable names and correlations

 @export
}
\description{
calculate pairwise correlations among variables using the BCD correlation
}
