% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_bart.R
\name{Predict_CRModel_BART}
\alias{Predict_CRModel_BART}
\title{Predict_CRModel_BART}
\usage{
Predict_CRModel_BART(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'CRModel_BART'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list with the following items.
CIFs:predicted CIF matrix (for event coded as 1),
Times: The times at which the CIFs are predicted.
}
\description{
Get predictions from a BART model for a test dataset
}
