% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRModelsFunctions.R
\name{Predict_CRModel_RF}
\alias{Predict_CRModel_RF}
\title{Predict_CRModel_RF}
\usage{
Predict_CRModel_RF(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_RF'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list with the following items. predSurvsTestRF: the output of 'randomForestSRC::predict.rfsrc',
 CIFs=predicted CIF matrix (for event coded as 1),
 Times: The times at which the CIFs are predicted.
}
\description{
Get predictions from a RF model for a test dataset
}
