% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRModelsFunctions.R
\name{Predict_CRModel_rulefit}
\alias{Predict_CRModel_rulefit}
\title{Predict_CRModel_rulefit}
\usage{
Predict_CRModel_rulefit(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'CRModel_rulefit'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list with the following items. TestMat: model matrix for the test data,
PredRFOut: prediction object from Fine-Gray model,
CIFs:predicted CIF matrix (for event coded as 1),
Times: The times at which the CIFs are predicted.
}
\description{
Get predictions from a rulefit model for a test dataset
}
