% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_bart.R
\name{Predict_SurvModel_BART}
\alias{Predict_SurvModel_BART}
\title{Predict_SurvModel_BART}
\usage{
Predict_SurvModel_BART(modelout, newdata, times = NULL)
}
\arguments{
\item{modelout}{the output from 'SurvModel_BART'}

\item{newdata}{the data for which the predictions are to be calculated}

\item{times}{optional vector of times for prediction (currently ignored by surv.pre.bart/predict).}
}
\value{
a list containing the following objects:
Probs: predicted survival probability matrix (rows=times, cols=observations),
Times: the times at which the probabilities are calculated (including 0, from model).
}
\description{
Make predictions using a fitted BART survival model.
}
