% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{Predict_SurvModel_Cox}
\alias{Predict_SurvModel_Cox}
\title{Predict_SurvModel_Cox}
\usage{
Predict_SurvModel_Cox(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_Cox'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
'cph_modelTestPredict': predictions as the output of 'survival::survfit'
}
\description{
Get predictions from a Cox model for a test dataset
}
