% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{Predict_SurvModel_deepsurv}
\alias{Predict_SurvModel_deepsurv}
\title{Predict_SurvModel_deepsurv}
\usage{
Predict_SurvModel_deepsurv(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_glmnet'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following objects. estSURVTest: predicted survival probability matrix
time.interest: times at which the survival probabilities are provided.
}
\description{
Get predictions from a deepsurv model for a test dataset
}
