% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gam.R
\name{Predict_SurvModel_GAM}
\alias{Predict_SurvModel_GAM}
\title{Predict_SurvModel_GAM}
\usage{
Predict_SurvModel_GAM(modelout, newdata, new_times = NULL)
}
\arguments{
\item{modelout}{the output from 'SurvModel_GAM' (a list containing 'model', 'times', 'varprof', 'expvars')}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following items:
 Probs: predicted Survival probability matrix (rows=times, cols=observations),
 Times: The times at which the probabilities are predicted.
}
\description{
Get predictions from a GAM survival model for a test dataset.
}
