% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gbm.R
\name{Predict_SurvModel_gbm}
\alias{Predict_SurvModel_gbm}
\title{Predict_SurvModel_gbm}
\usage{
Predict_SurvModel_gbm(modelout, newdata, new_times = NULL)
}
\arguments{
\item{modelout}{the output from 'SurvModel_gbm' (a list containing 'model', 'times', 'varprof', 'expvars')}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following items:
Probs: predicted survival probability matrix (rows=times, cols=observations),
Times: the unique times for which the probabilities are calculated (including 0).
}
\description{
Get predictions from a gbm survival model for a test dataset.
}
