% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_glmnet.R
\name{Predict_SurvModel_glmnet}
\alias{Predict_SurvModel_glmnet}
\title{Predict_SurvModel_glmnet}
\usage{
Predict_SurvModel_glmnet(modelout, newdata, newtimes = NULL)
}
\arguments{
\item{modelout}{the output from 'SurvModel_glmnet'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following items:
Probs: predicted survival probability matrix (rows=times, cols=observations),
Times: the unique times for which the probabilities are calculated (including 0).
}
\description{
Get predictions from a glmnet survival model for a test dataset.
}
