% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{Predict_SurvModel_SurvReg}
\alias{Predict_SurvModel_SurvReg}
\title{Predict_SurvModel_SurvReg}
\usage{
Predict_SurvModel_SurvReg(modelout, newdata, times = NULL)
}
\arguments{
\item{modelout}{the output from 'SurvModel_SurvReg'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following objects. estSURVTest: predicted survival probability matrix,
time.interest: the times at which the probabilities are calculated
}
\description{
Make predictions using BART model, trained by the
SurvModel_SurvReg' function
}
