% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{Predict_SurvModel_xgboost}
\alias{Predict_SurvModel_xgboost}
\title{Predict_SurvModel_xgboost}
\usage{
Predict_SurvModel_xgboost(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_xgboost'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following items: estSURVTest: survival probability matrix,
time.interest: the unique times for which the probabilities are calculated
}
\description{
Get predictions from a xgboost model for a test dataset
}
