% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecTools.R
\name{PredictAllPossibleOutcomesSurvOrCifs}
\alias{PredictAllPossibleOutcomesSurvOrCifs}
\title{PredictAllPossibleOutcomesSurvOrCifs}
\usage{
PredictAllPossibleOutcomesSurvOrCifs(data, modelslist, modeltypes, times)
}
\description{
Given a list of models and data get predictions
}
