% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ensemble.R
\name{PredictSurvModels}
\alias{PredictSurvModels}
\title{PredictSurvModels}
\usage{
PredictSurvModels(models, newdata, newtimes)
}
\arguments{
\item{models}{List containing the fitted models (output from 'RunSurvModels').}

\item{newdata}{Data frame with new data for prediction.}

\item{newtimes}{Numeric vector of times for which predictions are required.}
}
\value{
A list containing:
  - ModelPredictions: A list of individual model predictions (interpolated probability matrices).
  - NewProbs: Ensembled prediction matrix (averaged on cumulative hazard scale).
}
\description{
Utility function to get predictions from several survival models
at the same time and calculate an ensemble prediction.
}
