% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipesDataSteps.R
\name{prep_data_recipe}
\alias{prep_data_recipe}
\title{prep_data_recipe}
\usage{
prep_data_recipe(model_recipe, training)
}
\arguments{
\item{model_recipe}{initiated model recipe}

\item{training}{training data}
}
\value{
return a prepared recipe object
}
\description{
prepare a recipe using training data
}
