% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{prepare_stacking_data}
\alias{prepare_stacking_data}
\title{Helper to generate stacking predictions and actuals}
\usage{
prepare_stacking_data(task, fitted_models, time_grid, outcome_type, cal_data)
}
\description{
Helper to generate stacking predictions and actuals
}
\keyword{internal}
