% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_persistence.R
\name{print.CREnsemble}
\alias{print.CREnsemble}
\title{Print CREnsemble Object}
\usage{
\method{print}{CREnsemble}(x, ...)
}
\arguments{
\item{x}{CREnsemble object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print CREnsemble Object
}
