% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_persistence.R
\name{print.SurvEnsemble}
\alias{print.SurvEnsemble}
\title{Print SurvEnsemble Object}
\usage{
\method{print}{SurvEnsemble}(x, ...)
}
\arguments{
\item{x}{SurvEnsemble object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print a summary of a SurvEnsemble object
}
