% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{readDict}
\alias{readDict}
\title{readDict}
\usage{
readDict(file = NULL, filetype = "csv", ...)
}
\arguments{
\item{file}{path to dictionary file.}

\item{filetype}{type of dictionary file ("csv", "xlsx", "sas7bdat").}

\item{...}{additional arguments to pass to read.csv, openxlsx::read.xlsx,
haven::read_sas.}
}
\value{
data frame with dictionary information.
}
\description{
Import a dictionary that is in standard format.
The dictionary should be in a csv, xlsx, or sas7bdat file.
Standard columns: Variable, Description, Type, Value, Label, Notes.
}
\examples{
\dontrun{
dict <- readDict(file = "path/to/dictionary.csv", filetype = "csv")
}
}
