% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{RemoveAllNAVars}
\alias{RemoveAllNAVars}
\title{RemoveAllNAVars}
\usage{
RemoveAllNAVars(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame with all-NA columns removed.
}
\description{
Remove variables (columns) from a data frame that consist entirely of NA values.
}
