% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{RemoveEmptySpacesData}
\alias{RemoveEmptySpacesData}
\title{RemoveEmptySpacesData}
\usage{
RemoveEmptySpacesData(DATA)
}
\arguments{
\item{DATA}{data frame with data}
}
\value{
a data frame with data
}
\description{
Remove empty spaces from factors
}
