% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{RemoveMissinRecordsData}
\alias{RemoveMissinRecordsData}
\title{RemoveMissinRecordsData}
\usage{
RemoveMissinRecordsData(data, maxprop = 0.2)
}
\arguments{
\item{data}{data frame with data}

\item{maxprop}{numeric value between 0 and 1 for the maximum proportion of NA values}
}
\value{
data frame with data

 @export
}
\description{
Utility function to remove records with more than a certain percentage
of missing
}
