% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{RemoveMissinVarsData}
\alias{RemoveMissinVarsData}
\title{RemoveMissinVarsData}
\usage{
RemoveMissinVarsData(data, maxprop = 0.2)
}
\arguments{
\item{data}{data frame with data}

\item{maxprop}{numeric value between 0 and 1 for the maximum proportion of NA values}
}
\value{
data frame with data
}
\description{
Utility function to remove variables with more than a certain percentage
of missing
}
