% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{RemoveMonoVarsData}
\alias{RemoveMonoVarsData}
\title{RemoveMonoVarsData}
\usage{
RemoveMonoVarsData(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame with single-value columns removed.
}
\description{
Remove variables (columns) from a data frame that have only one unique non-NA value.
}
