% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{RemoveRareBinaryVarsData}
\alias{RemoveRareBinaryVarsData}
\title{RemoveRareBinaryVarsData}
\usage{
RemoveRareBinaryVarsData(data, minfreq = 0.01)
}
\arguments{
\item{data}{data frame with data}

\item{minfreq}{numeric value between 0 and 1 for the minimum frequency of a category}
}
\value{
data frame with data

 @export
}
\description{
Put NA's to rare (freq less than a specified percentage) categories of
binary variables
}
