% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_metrics.R
\name{restrictedMeanTimeLostCR}
\alias{restrictedMeanTimeLostCR}
\title{restrictedMeanTimeLostCR}
\usage{
restrictedMeanTimeLostCR(Predictions, times, UL, LL = 0)
}
\arguments{
\item{Predictions}{matrix of predicted CIFs (rows=times, cols=observations) for the cause of interest}

\item{times}{vector of time points corresponding to rows of Predictions}

\item{UL}{upper limit of integration}

\item{LL}{lower limit of integration (default: 0)}
}
\value{
vector of RMTL values for each observation
}
\description{
Calculate Restricted Mean Time Lost (RMTL) for competing risk predictions
}
