% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_steps.R
\name{rfimp_data_recipe}
\alias{rfimp_data_recipe}
\title{rfimp_data_recipe}
\usage{
rfimp_data_recipe(
  model_recipe,
  dummy = TRUE,
  onehot = FALSE,
  pmiss = 0.3,
  pother = 0.05,
  trees = 10,
  options = list(nbagg = 10, keepX = TRUE)
)
}
\arguments{
\item{model_recipe}{initiated model recipe}

\item{dummy}{use dummy coding or not}

\item{onehot}{whether onehot encoding is used or not (will miss one category)}

\item{pmiss}{percentage missing to delete a variable}

\item{pother}{percentage treshold to combine fator levels as other}

\item{trees}{number of trees to use in iputation}

\item{options}{a list of options to pass to the bagging algorith,
currently set as 'options = list(nbagg=10,keepX = TRUE)'}
}
\value{
return a  recipe object
}
\description{
Data recipe random forest imputation
}
