% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_ensemble.R
\name{RunCRModels}
\alias{RunCRModels}
\title{RunCRModels}
\usage{
RunCRModels(
  datatrain,
  ExpVars,
  timevar,
  eventvar,
  models = c("FG", "rulefit", "bart", "cox", "ttah"),
  ntreeRF = 300,
  varsel = FALSE,
  run_rf = TRUE
)
}
\arguments{
\item{datatrain}{data frame with explanatory and outcome variables}

\item{ExpVars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (coded as 0,1,2),
1 is the event of interest}

\item{models}{a vector of models to be fitted to be chosen among
"FG", "rulefit", "bart", "cox", "xgboost", "gam", "survreg", "ttah", "shallownn". Two random forest models are
automatically fitted and dont need to be listed here}

\item{ntreeRF}{number of trees for Random Forest models}

\item{varsel}{logical indicating whether variable selection to be
applied before fitting models "FG", "rulefit", "bart", "cox", "xgboost", "gam", "survreg"}

\item{run_rf}{logical; if `FALSE`, skips fitting the baseline Random Forest models.}
}
\value{
a list of two items. First is a list containing ExpVars: all of the explanatory vars,
 ExpVars2: a subset of the explanatory variables selected by random forest,
 timevar: time variable,
 eventvar: event variable,
 model_status: Named logical vector indicating which models succeeded (TRUE) or failed (FALSE).
 The second is also a list that contain the individual model outputs.
}
\description{
utility function to run several CR models at the same time
}
