% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{savedeepsurv}
\alias{savedeepsurv}
\title{savedeepsurv}
\usage{
savedeepsurv(modelout, filename)
}
\arguments{
\item{modelout}{the output from 'SurvModel_deepsurv'}

\item{filename}{the name of the file to save the model to}
}
\description{
Save a SurvModel_deepsurv model using ave_model_tf
}
