% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_persistence.R
\name{summary.CREnsemble}
\alias{summary.CREnsemble}
\title{Summary Method for CREnsemble}
\usage{
\method{summary}{CREnsemble}(object, ...)
}
\arguments{
\item{object}{CREnsemble object}

\item{...}{Additional arguments (ignored)}
}
\description{
Generate a summary of a CREnsemble object
}
