% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_persistence.R
\name{summary.SurvEnsemble}
\alias{summary.SurvEnsemble}
\title{Summary Method for SurvEnsemble}
\usage{
\method{summary}{SurvEnsemble}(object, ...)
}
\arguments{
\item{object}{SurvEnsemble object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for SurvEnsemble
}
