% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_persistence.R
\name{SurvEnsemble}
\alias{SurvEnsemble}
\title{Create SurvEnsemble S3 Class}
\usage{
SurvEnsemble(ensemble_list)
}
\arguments{
\item{ensemble_list}{List returned by RunSurvModels}
}
\value{
Object of class "SurvEnsemble"
}
\description{
Converts ensemble model output to S3 class for better organization
}
