% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_bart.R
\name{SurvModel_BART}
\alias{SurvModel_BART}
\title{SurvModel_BART}
\usage{
SurvModel_BART(data, expvars, timevar, eventvar, K = 8, ntree = 50)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}

\item{K}{parameter 'K' for BART (number of time points for discrete approximation)}

\item{ntree}{number of trees in BART model}
}
\value{
a list containing the following objects:
model: fitted BART model object from surv.bart,
times: time points used in the BART model,
varprof: profile of explanatory variables,
expvars: character vector of explanatory variables used,
factor_levels: list containing factor levels for consistent prediction.
}
\description{
Fit a BART model for survival outcomes using BART::surv.bart.
}
