% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_deepsurv}
\alias{SurvModel_deepsurv}
\title{SurvModel_deepsurv}
\usage{
SurvModel_deepsurv(
  data,
  expvars,
  timevar,
  eventvar,
  num_nodes = c(100L, 100L),
  dropout = 0.2,
  batch_size = 128L,
  epochs = 100L
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{num_nodes}{number of nodes in each of the hidden layers}

\item{dropout}{dropout percentage for each hiden layer}
}
\value{
a list contating the following objects:
expvars: character vector of explanatory variable names
learner: fitted learner
times: observed event times
}
\description{
Fit a deepsurv model for survival outcomes
 To use first do:

 library(survivalmodels)

 install_pycox(pip = TRUE, install_torch = TRUE)

 install_keras(pip = TRUE, install_tensorflow = TRUE)
}
\seealso{
[survivalmodels::deepsurv()] which this function wraps.
}
