% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gam.R
\name{SurvModel_GAM}
\alias{SurvModel_GAM}
\title{SurvModel_GAM}
\usage{
SurvModel_GAM(data, expvars, timevar, eventvar, shrinkTreshold = 10)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}

\item{shrinkTreshold}{integer value, minimum number of
factor levels for factor variables to be considered for shrinkage ('re' basis).}
}
\value{
a list with two elements. b: the fitted GAM model object from mgcv::gam,
model: list containing the internal Cox model ('model') and survfit object ('sf') from score2proba based on training predictions.
varprof: profile of explanatory variables.
}
\description{
Fit a GAM model for survival outcomes using Cox PH family.
}
