% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_gbm}
\alias{SurvModel_gbm}
\title{SurvModel_gbm}
\usage{
SurvModel_gbm(
  data,
  expvars,
  timevar,
  eventvar,
  ntree = 200,
  max.depth = 3,
  bag.fraction = 0.4,
  train.fraction = 0.4,
  learninrate = 0.01
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{ntree}{number of trees to grow}

\item{max.depth}{maximum depth for the trees}

\item{bag.fraction}{fraction of data to sample in each bagging iterations}

\item{train.fraction}{fraction for training data}

\item{learninrate}{learning rate for the boosting algorithm}
}
\value{
a list which contains the following objects.
gbmmodel: model object,
best.iter: best iteration number
expvars: vector of the names of explanatory variables,
survMat: survival probabilities,
basehaz.cum: baseline cumulative hazard
time.interest: unique observed event times
}
\description{
Fit a gbm model for survival outcomes using glmnet
}
