% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_glmnet.R
\name{SurvModel_glmnet}
\alias{SurvModel_glmnet}
\title{SurvModel_glmnet}
\usage{
SurvModel_glmnet(
  data,
  expvars,
  timevar,
  eventvar,
  alpha = 0.5,
  maxit = 5000,
  nfolds = 30
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}

\item{alpha}{elastic net mixing parameter (1=lasso, 0=ridge)}

\item{maxit}{maximum number of iterations for glmnet}

\item{nfolds}{number of folds for cross-validation in cv.glmnet}
}
\value{
a list containing the following objects:
traindata: a sample of original training predictors (for factor level consistency),
TrainMat: a sample of the model matrix used for training (for prediction structure),
yTrain: a sample of the Surv object used for training (for prediction structure),
expvars: character vector of explanatory variables used,
cv.fit: cv.glmnet fit object,
est.coef: estimated model coefficients at lambda.min,
varprof: profile of explanatory variables.
}
\description{
Fit a penalized Cox model for survival outcomes using glmnet.
}
