% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_random_forest.R
\name{SurvModel_RF}
\alias{SurvModel_RF}
\title{SurvModel_RF}
\usage{
SurvModel_RF(
  data,
  expvars,
  timevar,
  eventvar,
  ntree = 300,
  samplesize = 500,
  nsplit = 5,
  trace = TRUE
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}

\item{ntree}{integer value,  number of trees to grow}

\item{samplesize}{integer value,  sample size for each grown tree (swor)}

\item{nsplit}{integer value, maximum number of splits for each tree}

\item{trace}{logical, trace tuning process or not}
}
\value{
a list containing the fitted randomForestSRC object ('hd.obj') and
variable profile ('varprof').
}
\description{
Fit a RF model for survival outcomes using randomForestSRC.
Includes tuning of nodesize and mtry.
}
