% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_SurvReg}
\alias{SurvModel_SurvReg}
\title{SurvModel_SurvReg}
\usage{
SurvModel_SurvReg(data, expvars, timevar, eventvar, dist = "exponential")
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{dist}{one of "weibull", "exponential", "gaussian", "logistic","lognormal" or "loglogistic"}
}
\value{
a list containing the following objects. survregOut: survreg model out,
predicttrain: predictions in the training data,
times.interest: unique times for which the probabilities are calculated,
expvars: explanatory variables used to build the model
}
\description{
Fit parametric survival models
}
