% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_xgboost.R
\name{SurvModel_xgboost}
\alias{SurvModel_xgboost}
\title{SurvModel_xgboost}
\usage{
SurvModel_xgboost(data, expvars, timevar, eventvar)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}
}
\value{
a list containing the following objects:
expvars: vector of explanatory variables used,
learner: fitted model object (class xgb.Booster.surv),
estSURVTrain: predicted survival probabilities matrix for training data (rows=obs, cols=times),
datatrainProf: sample data frame of original predictors (for factor level consistency),
varprof: profile of explanatory variables.
}
\description{
Fit an xgboost model for survival outcomes.
}
