% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_steps.R
\name{t2edata_split}
\alias{t2edata_split}
\title{t2edata_split}
\usage{
t2edata_split(data, ...)
}
\arguments{
\item{data}{A dataset}

\item{...}{other parameters passed to 'rsample::initial_split'#'}
}
\value{
a list containing Train: training data and Test: test data
}
\description{
creates a list object with training and test dataset using
'rsample::initial_split'.
}
