% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_steps.R
\name{t2emodel_data_recipe_init}
\alias{t2emodel_data_recipe_init}
\title{t2emodel_data_recipe_init}
\usage{
t2emodel_data_recipe_init(timevar, eventvar, expvar, idvars, traindata)
}
\arguments{
\item{timevar}{character name of time variable}

\item{eventvar}{character name of event variable}

\item{expvar}{character vector of explanatory variable names}

\item{idvars}{character vector of id variable names}

\item{traindata}{training data}
}
\value{
return an initiated recipe object
}
\description{
initiate a model object using 'recipes::recipe'.
}
