% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{ZeroOneScalerData}
\alias{ZeroOneScalerData}
\title{ZeroOneScalerData}
\usage{
ZeroOneScalerData(data)
}
\arguments{
\item{data}{data frame with data}
}
\value{
a list with the data frame with data and the min and max values of each variable


 @export
}
\description{
Scale numeric variables to the range of 0 and 1
}
