% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{addDiseaseSpecColsDict}
\alias{addDiseaseSpecColsDict}
\title{addDiseaseSpecColsDict}
\usage{
addDiseaseSpecColsDict(dict)
}
\arguments{
\item{dict}{data frame with dictionary information}
}
\value{
data frame with dictionary information
 @export
}
\description{
process the dictionary to add disease specific / not applicable features
}
