% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{bindDicts}
\alias{bindDicts}
\title{bindDicts}
\usage{
bindDicts(Dict1, Dict2)
}
\arguments{
\item{Dict1}{data frame, the primary dictionary.}

\item{Dict2}{data frame, the secondary dictionary to append from.}
}
\value{
data frame containing combined dictionary information.
}
\description{
Utility function to combine two dictionaries. Rows from Dict2 for variables
already present in Dict1 are ignored.
}
