% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{BrierScore}
\alias{BrierScore}
\title{BrierScore}
\usage{
BrierScore(predsurv, predsurvtimes, obstimes, obsevents)
}
\arguments{
\item{predsurv}{Predicted matrix of probabilities}

\item{predsurvtimes}{The times for which the probabilities are predicted}

\item{obstimes}{Observed times}

\item{obsevents}{Observed event indicator}
}
\value{
the output from the 'pec::pec' function
}
\description{
get time dependent brier score (PEC)
}
