% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_metrics.R
\name{BrierScore}
\alias{BrierScore}
\title{BrierScore}
\usage{
BrierScore(predsurv, predsurvtimes, obstimes, obsevents)
}
\arguments{
\item{predsurv}{Predicted matrix of survival probabilities (rows=times, cols=observations).}

\item{predsurvtimes}{The times for which the probabilities are predicted.}

\item{obstimes}{Observed times vector.}

\item{obsevents}{Observed event indicator vector (0=censored, 1=event).}
}
\value{
The output object from the 'pec::pec' function, containing Brier scores over time.
}
\description{
Calculate the time-dependent Brier score using pec::pec.
}
