% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_metrics.R
\name{BrierScoreCR}
\alias{BrierScoreCR}
\title{BrierScoreCR}
\usage{
BrierScoreCR(SurvObj, Predictions, time, cause = 1, TestMat = NULL)
}
\arguments{
\item{SurvObj}{survival object (Surv(time, status))}

\item{Predictions}{matrix of predicted CIFs (rows=observations, cols=times) for the cause of interest}

\item{time}{evaluation time point}

\item{cause}{cause of interest (1 or 2)}

\item{TestMat}{test dataset (optional, used if formula needs predictors)}
}
\value{
Brier score value
}
\description{
Calculate Brier score for competing risk predictions at specific times
}
