% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{CollapseRareCategoriestoOtherData}
\alias{CollapseRareCategoriestoOtherData}
\title{CollapseRareCategoriestoOtherData}
\usage{
CollapseRareCategoriestoOtherData(data, minfreq = 0.01)
}
\arguments{
\item{data}{data frame with data}

\item{minfreq}{numeric value between 0 and 1 for the minimum frequency of a category}
}
\value{
data frame with data

 @export
}
\description{
Put 'Other' to rare (freq less than a specified percentage) categories of
factor (Categorical) variables
}
