% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{ConVarsDict}
\alias{ConVarsDict}
\title{ConVarsDict}
\usage{
ConVarsDict(dict)
}
\arguments{
\item{dict}{data frame with dictionary information}
}
\value{
data frame with dictionary information
}
\description{
Utility function to get dictionary for the continuous variables
}
