% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRModelsFunctions.R
\name{CRModel_BART}
\alias{CRModel_BART}
\title{CRModel_BART}
\usage{
CRModel_BART(data, expvars, timevar, eventvar, K = 10, ntree = 100)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{K}{number of cutpoints for the time variable}

\item{ntree}{number of trees to fit to extract rules}
}
\value{
a list object with the following: post: BART model,
 expvars: character vector of explanatory variables,
 eventvarvec: event variable values,timevarvec: time variable values,
 x.train: design matrix for the training data,
 varprof: profile of the explanatory variables
}
\description{
Fit a BART model for competing risks outcomes
}
