% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRModelsFunctions.R
\name{CRModel_Cox}
\alias{CRModel_Cox}
\title{CRModel_Cox}
\usage{
CRModel_Cox(data, expvars, timevar, eventvar)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (coded 0,1,2)}
}
\value{
a list object with the following. cph_model: trained Cox PH model object,
varprof: profile of explanatory variables.
}
\description{
Fit a Cox model for CR outcomes using forward selection
}
