% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRModelsFunctions.R
\name{CRModel_FineGray}
\alias{CRModel_FineGray}
\title{CRModel_FineGray}
\usage{
CRModel_FineGray(data, expvars, timevar, eventvar)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (coded 0,1,2)}
}
\value{
a list object with the following. allvals: AIC values for a range of shrinkage parameters,
fit: fit object from 'fastcmprsk::fastCrrp',
CIF.hat: estimated CIFs in training data,
expvars: character vector of explanatory variables,
timesfit: times for which CIFs are given,
varprof: profile of explanatory variables.
}
\description{
Fit a Fine-Gray model for CR outcomes using penalized regression
}
