% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_random_forest.R
\name{CRModel_RF}
\alias{CRModel_RF}
\title{CRModel_RF}
\usage{
CRModel_RF(
  data,
  expvars,
  timevar,
  eventvar,
  cause = c(1, 1),
  ntree = 300,
  samplesize = 500,
  nsplit = 5,
  trace = TRUE
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (coded 0,1,2)}

\item{cause}{vectoe of weights for causes}

\item{ntree}{number of trees to fit to extract rules}

\item{samplesize}{number of samples for each tree}

\item{nsplit}{maximum number of splits for each tree}

\item{trace}{trace the results or not}
}
\value{
a list object with the following. hd.obj: trained random forest model object,
varprof: profile of explanatory variables.
}
\description{
Fit a  random forest model for competing risk outcomes
}
