% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_rulefit.R
\name{CRModel_rulefit}
\alias{CRModel_rulefit}
\title{CRModel_rulefit}
\usage{
CRModel_rulefit(
  data,
  expvars,
  timevar,
  eventvar,
  ntree = 300,
  nsample = 300,
  keepvars = NULL,
  cuttimes = NULL
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{ntree}{number of trees to fit to extract rules}

\item{nsample}{number of samples for each tree}

\item{keepvars}{these variables will be used in each bagging iteration}

\item{cuttimes}{these cut times are used for calculating pseudo observation}
}
\value{
a list object with the following: datasamp: a sample of data,
ruleslist: list of the fitted rules,
ctreelist: list of tree models,
CRrulefitModel: fitted Fine-Gray model object (used internally by rulefit),
expvars: explanatory variables used in the model
usecols: columns used in the internal Fine-Gray model
varprof: profile of explanatory variables.
}
\description{
Fit a rulefit model for competing risks outcomes
}
