% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{DateVarstoDateDict}
\alias{DateVarstoDateDict}
\title{DateVarstoDateDict}
\usage{
DateVarstoDateDict(dict, datevars = NULL)
}
\arguments{
\item{dict}{data frame with dictionary information.}

\item{datevars}{character vector of variable names to mark as "Date".}
}
\value{
data frame with updated 'Type' for date variables.
}
\description{
Utility function to change the 'Type' of specified date variables to "Date".
}
