% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{DetectNAStringsDict}
\alias{DetectNAStringsDict}
\title{NumIdVarstoCharDict}
\usage{
DetectNAStringsDict(dict, knownNaLabels = NULL, knownNaValues = NULL)
}
\arguments{
\item{dict}{data frame with dictionary information}

\item{knownNaLabels}{character vector of known NA labels}

\item{knownNaValues}{character vector of known NA values}
}
\value{
data frame with dictionary information
}
\description{
Utility function to detect NA strings in a dictionary
}
