% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{droplevelsoffactorsData}
\alias{droplevelsoffactorsData}
\title{droplevelsoffactorsData}
\usage{
droplevelsoffactorsData(data)
}
\arguments{
\item{data}{data frame with data}
}
\value{
data frame with data


 @export
}
\description{
Drop unused factor levels from data
}
