% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{filllasttoDict}
\alias{filllasttoDict}
\title{filllasttoDict}
\usage{
filllasttoDict(dict)
}
\arguments{
\item{dict}{data frame with dictionary information.}
}
\value{
data frame with NA values filled forward column-wise.
}
\description{
Utility function to fill missing values (NA) in dictionary columns
by carrying forward the last non-NA value.
}
