% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_detection.R
\name{finddatevars}
\alias{finddatevars}
\title{finddatevars}
\usage{
finddatevars(names, knowndatevars = NULL)
}
\arguments{
\item{names}{character vector of variable names.}

\item{knowndatevars}{Optional character vector of known date variable names to include.}
}
\value{
character vector of potential date variable names found in 'names'.
}
\description{
Utility function to find potential date variables based on common patterns in names.
}
