% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{findDiseaseSpecVarsData}
\alias{findDiseaseSpecVarsData}
\title{findDiseaseSpecVarsData}
\usage{
findDiseaseSpecVarsData(data, pin = 0.5, pout = 0.9)
}
\arguments{
\item{data}{data frame with data}

\item{pin}{numeric value between 0 and 1 for the proportion of non NA values in the disease group}

\item{pout}{numeric value between 0 and 1 for the proportion of NA values in the other groups}
}
\value{
list of variables that are disease specific

 @export
}
\description{
Utility function to try to detect disease spec variables from data
}
