% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{getcharcolsData}
\alias{getcharcolsData}
\title{getcharcolsData}
\usage{
getcharcolsData(data)
}
\arguments{
\item{data}{data frame with data}
}
\value{
character vector with names of character columns

 @export
}
\description{
Utility function to det the names of Character columns in data
}
