% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{gethighcorvarsDataSummmary}
\alias{gethighcorvarsDataSummmary}
\title{gethighcorvarsDataSummmary}
\usage{
gethighcorvarsDataSummmary(pmat, corcutoff = 0.8)
}
\arguments{
\item{pmat}{correlation matrix}

\item{corcutoff}{correlation cutoff}
}
\value{
a matrix with variable names and correlations


 @export
}
\description{
Filter correlations based on a threshold and list names
}
