% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{ImputeMissinRecordsData}
\alias{ImputeMissinRecordsData}
\title{ImputeMissinRecordsData}
\usage{
ImputeMissinRecordsData(data, dontuse = NULL, ...)
}
\arguments{
\item{data}{data set to be imputed}

\item{dontuse}{the variables not to use when imputing,
thesev ariables will not be used for imputation or for imputing other
 variables.}

\item{...}{Additional arguments passed to \code{\link[ranger:ranger]{ranger::ranger()}}. Not all make sense.}
}
\value{
Imputed data
}
\description{
Imputation of data using missRanger
}
\examples{
\dontrun{
DATAImputed<-DATA\%>\%ImputeMissinRecordsData()
}

}
\seealso{
[missRanger::missRanger()] which this function wraps.
}
