% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{integratedBrier}
\alias{integratedBrier}
\title{integratedBrier}
\usage{
integratedBrier(times, scores, minmax = c(1, 35))
}
\arguments{
\item{times}{The times at which the concordance scores were calculated}

\item{scores}{Time dependent Brier scores}

\item{minmax}{The min max limits of the integral}
}
\value{
a scalar measure of accuracy
}
\description{
Integrated Brier score
}
