% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_metrics.R
\name{integratedBrier}
\alias{integratedBrier}
\title{integratedBrier}
\usage{
integratedBrier(times, scores, minmax = c(1, 35))
}
\arguments{
\item{times}{The times at which the Brier scores were calculated.}

\item{scores}{Time-dependent Brier scores corresponding to 'times'.}

\item{minmax}{Numeric vector of length 2 specifying the integration limits [min, max].}
}
\value{
A scalar value representing the integrated Brier score, scaled by the interval length.
}
\description{
Calculate the integrated Brier score (IBS) over a specified time range.
}
