% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_metrics.R
\name{integratedC}
\alias{integratedC}
\title{integratedC}
\usage{
integratedC(times, scores, minmax = c(1, 35))
}
\arguments{
\item{times}{The times at which the C-index scores were calculated.}

\item{scores}{Time-dependent C-index scores corresponding to 'times'.}

\item{minmax}{Numeric vector of length 2 specifying the integration limits [min, max].}
}
\value{
A scalar value representing the integrated C-index, scaled by the interval length.
}
\description{
Calculate the integrated C-index over a specified time range.
}
