% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{MakelabelledSASDict}
\alias{MakelabelledSASDict}
\title{MakelabelledSASDict}
\usage{
MakelabelledSASDict(fileloc)
}
\arguments{
\item{fileloc}{Path to the SAS dataset file (.sas7bdat).}
}
\value{
A data frame representing the dictionary derived from SAS labels.
}
\description{
Create a dictionary data frame from a labelled SAS dataset file.
Uses the 'labelled' package to generate the dictionary.
}
