% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{MakeTestDataConfWithTrainData}
\alias{MakeTestDataConfWithTrainData}
\title{MakeTestDataConfWithTrainData}
\usage{
MakeTestDataConfWithTrainData(traindata, testdata)
}
\arguments{
\item{traindata}{data frame with training data}

\item{testdata}{data frame with test data}
}
\value{
a data frame with test data


 @export
}
\description{
Make sure test data does not have levels that are not observed in the training
}
