% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{ml4t2e_calibrate}
\alias{ml4t2e_calibrate}
\title{Calibrate a fitted model for conformal prediction}
\usage{
ml4t2e_calibrate(object, data)
}
\arguments{
\item{object}{A \code{t2e_fit} object.}

\item{data}{A data frame to be used for calibration. Must contain the same time/event columns as the training data.}
}
\value{
A \code{t2e_fit} object with a \verb{$conformal} slot containing scores and weights.
}
\description{
Computes conformal scores on a calibration set and attaches them to the fit object.
}
