% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fit.R, R/persistence.R
\name{ml4t2e_load}
\alias{ml4t2e_load}
\title{Load a persisted fit}
\usage{
ml4t2e_load(file)

ml4t2e_load(file)
}
\arguments{
\item{file}{Path to the saved object.}

\item{path}{Path to a saved fit (created by `ml4t2e_save()`).}
}
\value{
A `t2e_fit` object.

The loaded `t2e_fit` or `ml4t2e_pipeline` object.
}
\description{
Loads a previously saved `ml4time2event` object.
}
