% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fit.R
\name{ml4t2e_load}
\alias{ml4t2e_load}
\title{Load a persisted fit}
\usage{
ml4t2e_load(path)
}
\arguments{
\item{path}{Path to a saved fit (created by `ml4t2e_save()`).}
}
\value{
A `t2e_fit` object.
}
\description{
Load a persisted fit
}
