% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{ml4t2e_predict_censoring}
\alias{ml4t2e_predict_censoring}
\title{Predict Censoring Probabilities}
\usage{
ml4t2e_predict_censoring(model, times)
}
\arguments{
\item{model}{A `survfit` object (censoring model).}

\item{times}{Numeric vector of times to predict.}
}
\value{
Numeric vector of probabilities P(C > t).
}
\description{
Predict Censoring Probabilities
}
\keyword{internal}
