% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_fit.R
\name{ml4t2e_save}
\alias{ml4t2e_save}
\title{Persist fitted objects to disk}
\usage{
ml4t2e_save(fit, path)
}
\arguments{
\item{fit}{A `t2e_fit` object.}

\item{path}{Path to the file to create.}
}
\description{
Persist fitted objects to disk
}
