% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_processing.R
\name{NumIdVarstoCharDict}
\alias{NumIdVarstoCharDict}
\title{NumIdVarstoCharDict}
\usage{
NumIdVarstoCharDict(dict, idvars = c("`#`", "pseudoid", "pseudoccn"))
}
\arguments{
\item{dict}{data frame with dictionary information.}

\item{idvars}{character vector of variable names to treat as character IDs.}
}
\value{
data frame with updated 'Type' for ID variables.
}
\description{
Utility function to change the 'Type' of specified ID variables to "Character".
}
