% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{NumVarstCatsData}
\alias{NumVarstCatsData}
\title{NumVarstCatsData}
\usage{
NumVarstCatsData(data, numgroups = NULL, cuts = NULL)
}
\arguments{
\item{data}{data frame with data}

\item{numgroups}{a vector with the number of groups for each variable}

\item{cuts}{a vector with the cut points for each variable}
}
\value{
a data frame with the data


 @export
}
\description{
Make categorcal variables from numeric variables
}
