% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{OneAgainstRestCorDataSummmary}
\alias{OneAgainstRestCorDataSummmary}
\title{OneAgainstRestCorDataSummmary}
\usage{
OneAgainstRestCorDataSummmary(data)
}
\arguments{
\item{data}{Data frame. Factors/characters are converted to dummy variables.}
}
\value{
A named numeric vector where names are the original variable names (or derived numeric feature names)
  and values are the distance correlations of each against the rest.
}
\description{
Calculate the distance correlation of each variable against all other variables combined.
Note: This can be computationally intensive.
}
