% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{OneValCat2NumDict}
\alias{OneValCat2NumDict}
\title{OneValCat2NumDict}
\usage{
OneValCat2NumDict(dict, vars)
}
\arguments{
\item{dict}{data frame with dictionary information}

\item{vars}{character vector of variable names}
}
\value{
data frame with dictionary information
}
\description{
Utility function to correct the class of Continuous variables if they were
misidentified
}
