% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_fine_gray.R
\name{Predict_CRModel_FG}
\alias{Predict_CRModel_FG}
\title{Predict_CRModel_FG}
\usage{
Predict_CRModel_FG(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'CRModel_FineGray'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list with the following items. CIFs=predicted CIF matrix (for event coded as 1),
 Times: The times at which the CIFs are predicted.
}
\description{
Get predictions from a Fine-Gray CR model for a test dataset
}
