% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_deepsurv.R
\name{predict_risk_scores}
\alias{predict_risk_scores}
\title{(Helper) Compute Risk Scores}
\usage{
predict_risk_scores(modelout, x_matrix)
}
\arguments{
\item{modelout}{Fitted DeepSurv model object.}

\item{x_matrix}{Model matrix produced by `prepare_newdata_for_model`.}
}
\value{
Numeric vector of risk scores exp(η(x)).
}
\description{
Generates exponentiated linear predictors for a prepared design matrix.
}
\keyword{internal}
