% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{Predict_SurvModel_BART}
\alias{Predict_SurvModel_BART}
\title{Predict_SurvModel_BART}
\usage{
Predict_SurvModel_BART(modelout, newdata, times = NULL)
}
\arguments{
\item{modelout}{the output from 'SurvModel_BART'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following objects. PredMat: predicted survival probability matrix,
times: the times at which the probabilities are calculated.
}
\description{
Make predictions using BART model, trained by the 'SurvModel_BART' function
}
