% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_cox.R
\name{Predict_SurvModel_Cox}
\alias{Predict_SurvModel_Cox}
\title{Predict_SurvModel_Cox}
\usage{
Predict_SurvModel_Cox(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_Cox' (list containing 'cph_model' and 'times')}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing:
Probs: predicted survival probability matrix (rows=times, cols=observations),
Times: the times at which the probabilities are calculated (including 0),
cph_modelTestPredict: the raw output object from 'survival::survfit'.
}
\description{
Get predictions from a Cox survival model for a test dataset.
}
