% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gam.R
\name{Predict_SurvModel_GAM}
\alias{Predict_SurvModel_GAM}
\title{Predict_SurvModel_GAM}
\usage{
Predict_SurvModel_GAM(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_GAM'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list object with three items. Probs: survival probability predictions matrix (rows=times, cols=observations),
Times: the times of the returned probabilities,
sf: prediction object from survival::survfit.
}
\description{
Get predictions from a GAM survival model for a test dataset.
}
