% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{Predict_SurvModel_glmnet}
\alias{Predict_SurvModel_glmnet}
\title{Predict_SurvModel_glmnet}
\usage{
Predict_SurvModel_glmnet(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_glmnet'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following items: predSurvsTest: survival probability matrix,
timesTest: the unique times for which the probabilities are calculated
}
\description{
Get predictions from a glmnet model for a test dataset
}
