% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{Predict_SurvModel_RF}
\alias{Predict_SurvModel_RF}
\title{Predict_SurvModel_RF}
\usage{
Predict_SurvModel_RF(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_RF'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following items:
 Probs=predicted Survival probability matrix,
 Times: The times at which the CIFs are predicted,
 predSurvsTestRF:the output of 'randomForestSRC::predict.rfsrc'.
}
\description{
Get predictions from a RF model for a test dataset
}
