% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_random_forest.R
\name{Predict_SurvModel_RF}
\alias{Predict_SurvModel_RF}
\title{Predict_SurvModel_RF}
\usage{
Predict_SurvModel_RF(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_RF' (a list containing 'hd.obj')}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list containing the following items:
 Probs: predicted Survival probability matrix (rows=times, cols=observations),
 Times: The times at which the probabilities are predicted,
 predSurvsTestRF: the raw output object from 'randomForestSRC::predict.rfsrc'.
}
\description{
Get predictions from a RF survival model for a test dataset.
}
