% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_rulefit.R
\name{Predict_SurvModel_rulefit}
\alias{Predict_SurvModel_rulefit}
\title{Predict_SurvModel_rulefit}
\usage{
Predict_SurvModel_rulefit(modelout, newdata)
}
\arguments{
\item{modelout}{the output from 'SurvModel_rulefit'}

\item{newdata}{the data for which the predictions are to be calculated}
}
\value{
a list object with three items. Probs: survival probability predictions matrix (rows=times, cols=observations),
Times: the times of the returned probabilities, sfitTestRules: prediction object from survfit
}
\description{
Get predictions from a rulefit model for a test dataset
}
