% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_rulefit.R
\name{Predict_SurvModel_rulefit}
\alias{Predict_SurvModel_rulefit}
\title{Predict Survival Probabilities from a RuleFit Model}
\usage{
Predict_SurvModel_rulefit(modelout, newdata, new_times = NULL)
}
\arguments{
\item{modelout}{The output from `SurvModel_rulefit` (a list containing 'model', 'times', 'varprof', 'expvars', 'factor_levels').}

\item{newdata}{A new data frame for which to generate predictions. It must contain
all variables specified in `modelout$expvars`.}
}
\value{
A list with two items:
\item{Probs}{A matrix of survival probability predictions, where rows are time points and columns are observations.}
\item{Times}{The vector of time points corresponding to the rows of `Probs`.}
}
\description{
Generates survival probability predictions for new data using a
previously trained RuleFit model.
}
