% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_reg.R
\name{Predict_SurvModel_SurvReg}
\alias{Predict_SurvModel_SurvReg}
\title{Predict_SurvModel_SurvReg}
\usage{
Predict_SurvModel_SurvReg(modelout, newdata, new_times = NULL)
}
\arguments{
\item{modelout}{the output from 'SurvModel_SurvReg'}

\item{newdata}{the data for which the predictions are to be calculated}

\item{times}{optional vector of times for prediction. If NULL, uses times from modelout.}
}
\value{
a list containing the following objects:
Probs: predicted survival probability matrix (rows=times, cols=observations),
Times: the times at which the probabilities are calculated (including 0).
}
\description{
Make predictions using a fitted parametric survival model.
}
