% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRModelsFunctions.R
\name{PredictCRModels}
\alias{PredictCRModels}
\title{PredictCRodels}
\usage{
PredictCRModels(models, newdata, newtimes)
}
\arguments{
\item{models}{the output from 'RunCRModels'}

\item{newdata}{the data for which the predictions are to be calculated}

\item{newtimes}{the times for which the predictions obtained}
}
\value{
a list containing the following objects.
ModelPredictions: a list of individual model predictions,
NewProbs: ensembled predictions for CIFs
}
\description{
Utility function to get predictions from  several competing risks models
at the same time. Also adds an ensemble prediction for the probabilities.
}
