% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{processDiseaseSpecVarsData}
\alias{processDiseaseSpecVarsData}
\title{processDiseaseSpecVarsData}
\usage{
processDiseaseSpecVarsData(data, dislist)
}
\arguments{
\item{data}{data frame with data}

\item{dislist}{list of variables that are disease specific}
}
\value{
data frame with data
}
\description{
Utility function to process disease spec variables
}
