% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{readData}
\alias{readData}
\title{readData}
\usage{
readData(file = NULL, filetype = "csv", ...)
}
\arguments{
\item{file}{character vector of file location}

\item{filetype}{character vector of file type (csv, xlsx, sas7bdat)}
}
\value{
data frame with data

 @export
}
\description{
Import a data that is in standard format
}
