% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{readDict}
\alias{readDict}
\title{readDict}
\usage{
readDict(file = NULL, filetype = "csv", ...)
}
\arguments{
\item{file}{path to dictionary file}

\item{filetype}{type of dictionary file (csv, xlsx, sas7bdat)}

\item{...}{additional arguments to pass to read.csv, openxlsx::read.xlsx,
haven::read_sas}
}
\value{
data frame with variable information
}
\description{
Import a dictionary that is in standard format
The dictionary should be in a csv, xlsx, or sas7bdat file
}
\examples{
\dontrun{
readDict(file = "data/dictionary.csv", filetype = "csv")
}
}
