% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{RemoveEmptySpacesData}
\alias{RemoveEmptySpacesData}
\title{RemoveEmptySpacesData}
\usage{
RemoveEmptySpacesData(DATA)
}
\arguments{
\item{DATA}{Data frame.}
}
\value{
Data frame with whitespace trimmed.
}
\description{
Remove leading/trailing whitespace from character columns and factor levels.
}
