% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{RemoveMonoVarsData}
\alias{RemoveMonoVarsData}
\title{RemoveMonoVarsData}
\usage{
RemoveMonoVarsData(data)
}
\arguments{
\item{data}{data frame with data}
}
\value{
data frame with data
}
\description{
If a variable has one value remove it
}
