% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{ReplaceOutlierNumValsData}
\alias{ReplaceOutlierNumValsData}
\title{ReplaceOutlierNumValsData}
\usage{
ReplaceOutlierNumValsData(data, multIQR = 1.5, minnumgroup = 10)
}
\arguments{
\item{data}{data frame with data}

\item{multIQR}{multiplier for IQR}

\item{minnumgroup}{minimum number of groups to consider a variable}
}
\value{
a data frame with outliers replaced



 @export
}
\description{
Replace the outlier values for numeric vars
}
