% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{RunSurvModels}
\alias{RunSurvModels}
\title{RunSurvModels}
\usage{
RunSurvModels(
  datatrain,
  ExpVars,
  timevar,
  eventvar,
  models = c("glmnet", "coxph", "rulefit", "xgboost", "deepsurv", "gam", "gbm",
    "ExpSurvReg", "WeibSurvReg", "bart"),
  ntreeRF = 300,
  nvars = 20,
  ...
)
}
\arguments{
\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{models}{a vector of models to be fitted to be chosen among
"glmnet","coxph","rulefit","xgboost","deepsurv","gam","gbm",
"ExpSurvReg","WeibSurvReg","bart". Two random forest models are
automatically fitted and dont need to be listed here}

\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}
}
\value{
a list of two items. First is a list containing ExpVars: all of the explanatory vars,
 ExpVars2: a subset of the explanatory variables selected by random forest,
 timevar: time variable,
 eventvar: event variable. The second is also a list that contain the individual model outputs.
}
\description{
utility function to run several survival models at the same time
}
