% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{SummaryTableDataSummmary}
\alias{SummaryTableDataSummmary}
\title{SummaryTableDataSummmary}
\usage{
SummaryTableDataSummmary(data, UseVars)
}
\arguments{
\item{data}{data frame with data}

\item{UseVars}{a vector of variables to include in the summary}
}
\value{
a gtsummary table


 @export
}
\description{
Make a data summary of a list of variables in the data
}
