% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_BART}
\alias{SurvModel_BART}
\title{SurvModel_BART}
\usage{
SurvModel_BART(data, expvars, timevar, eventvar, K = 8, ntree = 50)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{K}{parameter 'K' for BART}
}
\description{
Fit a BART model
}
