% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_cox.R
\name{SurvModel_Cox}
\alias{SurvModel_Cox}
\title{SurvModel_Cox}
\usage{
SurvModel_Cox(data, expvars, timevar, eventvar)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}
}
\value{
a list of three items: cph_model: the final Cox PH model object after step selection,
 times: unique event times from the training data,
 varprof: profile of explanatory variables.
}
\description{
Fit a Cox model for survival outcomes using forward selection via stats::step.
}
