% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_Cox}
\alias{SurvModel_Cox}
\title{SurvModel_Cox}
\usage{
SurvModel_Cox(data, expvars, timevar, eventvar)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}
}
\value{
a list of two items. cph_model: coxPH model object,
 times: unique times in the data
}
\description{
Fit a Cox model for survival outcomes using forward selection
}
