% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_deepsurv.R
\name{SurvModel_DeepSurv}
\alias{SurvModel_DeepSurv}
\title{SurvModel_DeepSurv}
\usage{
SurvModel_DeepSurv(
  data,
  expvars,
  timevar,
  eventvar,
  size = 5,
  decay = 0.01,
  maxit = 1000,
  verbose = FALSE
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}

\item{size}{integer, number of units in the hidden layer (default: 5)}

\item{decay}{numeric, L2 regularization parameter (default: 0.01)}

\item{maxit}{integer, maximum iterations for optimization (default: 1000)}

\item{verbose}{logical, print progress messages (default: FALSE)}
}
\value{
a list with the following components:
  \item{model}{fitted neural network model (list with weights)}
  \item{times}{unique event times from training data}
  \item{varprof}{variable profile list}
  \item{expvars}{character vector of explanatory variables}
  \item{factor_levels}{list of factor levels for categorical variables}
}
\description{
Fit a DeepSurv neural network model for survival outcomes using a custom implementation with Cox loss.
}
