% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_GAM}
\alias{SurvModel_GAM}
\title{SurvModel_GAM}
\usage{
SurvModel_GAM(data, expvars, timevar, eventvar, shrinkTreshold = 10)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{shrinkTreshold}{integer value, minimum number of
factor levels for factor variables to be shrunk}
}
\value{
a list with two elements. b: the fitted GAM model,
model: predictions and related output
}
\description{
Fit a GAM model for survival outcomes
}
