% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gbm.R
\name{SurvModel_gbm}
\alias{SurvModel_gbm}
\title{SurvModel_gbm}
\usage{
SurvModel_gbm(
  data,
  expvars,
  timevar,
  eventvar,
  ntree = 200,
  max.depth = 3,
  bag.fraction = 0.3,
  train.fraction = 0.3,
  learninrate = 0.01
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}

\item{ntree}{number of trees to grow}

\item{max.depth}{maximum depth for the trees}

\item{bag.fraction}{fraction of data to sample in each bagging iteration}

\item{train.fraction}{fraction for training data (used internally by gbm for CV/OOB error)}

\item{learninrate}{learning rate (shrinkage) for the boosting algorithm}
}
\value{
a list of four items: model: fitted gbm model object,
 times: unique event times from the training data,
 varprof: profile of explanatory variables,
 expvars: the explanatory variables used.
}
\description{
Fit a gbm model for survival outcomes using gbm package.
}
