% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_glmnet}
\alias{SurvModel_glmnet}
\title{SurvModel_glmnet}
\usage{
SurvModel_glmnet(
  data,
  expvars,
  timevar,
  eventvar,
  alpha = 0.5,
  maxit = 5000,
  nfolds = 30
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{alpha}{lasso: alpha = 1; ridge: alpha=0}

\item{maxit}{maximum number of iterations for glmnet}
}
\value{
a list containing the following objects. traindata: a sample of training data,
TrainMat: preprocessed version of training data,
yTrain: a sample for response variable,
expvars: a character vector of explanatory variables,
cv.fit: glmnet cv fit object,
est.coef: estimated model coefficients
}
\description{
Fit a Cox model for survival outcomes using glmnet
}
