% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_random_forest.R
\name{SurvModel_RF}
\alias{SurvModel_RF}
\title{SurvModel_RF}
\usage{
SurvModel_RF(
  data,
  expvars,
  timevar,
  eventvar,
  ntree = 300,
  samplesize = 500,
  nsplit = 5,
  trace = TRUE,
  splitrule = "bs.gradient",
  nodesize_try = c(1, 5, 10, 15),
  importance = "permute",
  ...
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}

\item{ntree}{integer value, number of trees to grow (default: 300)}

\item{samplesize}{integer value, sample size for each grown tree (default: 500)}

\item{nsplit}{integer value, maximum number of splits for each tree (default: 5)}

\item{trace}{logical, trace tuning process or not (default: TRUE)}

\item{splitrule}{character, split rule for trees (default: "bs.gradient")}

\item{nodesize_try}{numeric vector, nodesize values to try during tuning (default: c(1, 5, 10, 15))}

\item{...}{additional parameters passed to randomForestSRC functions}
}
\value{
a list of four items: model: fitted randomForestSRC model object,
 times: unique event times from the training data,
 varprof: profile of explanatory variables,
 expvars: the explanatory variables used.
}
\description{
Fit a RF model for survival outcomes using randomForestSRC.
Includes tuning of nodesize and mtry.
}
