% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_RF}
\alias{SurvModel_RF}
\title{SurvModel_RF}
\usage{
SurvModel_RF(
  data,
  expvars,
  timevar,
  eventvar,
  ntree = 300,
  samplesize = 500,
  nsplit = 5,
  trace = TRUE
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{ntree}{integer value,  number of trees to grow}

\item{samplesize}{integer value,  sample size for each grown tree}

\item{nsplit}{integer value, maximum number of splits for each tree}

\item{trace}{trace or not, logical}
}
\value{
the output of 'randomForestSRC::rfsrc'
}
\description{
Fit a RF model for survival outcomes
}
