% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_rulefit.R
\name{SurvModel_rulefit}
\alias{SurvModel_rulefit}
\title{SurvModel_rulefit}
\usage{
SurvModel_rulefit(
  data,
  expvars,
  timevar,
  eventvar,
  ntree = 300,
  nsample = 300,
  keepvars = NULL,
  cuttimes = NULL
)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}

\item{ntree}{number of trees to fit to extract rules}

\item{nsample}{number of samples for each tree}

\item{keepvars}{these variables will be used in each bagging iteration}

\item{cuttimes}{these cut times are used for calculating pseudo observation}
}
\value{
a list object with the following: datasamp: a sample of data,
varprof: profile of explanatory variables,
ruleslist: list of the fitted rules,
TrainMat: fitted rules matrix (combined with original features),
ctreelist: list of tree models,
yTrain: survival object for outcome variables,
cv.fitRules: fitted glmnet survival models for a range of shrinkage parameters,
timesTrain: unique times in the outcome variable from survfit object
expvars: explanatory variables used in the model
}
\description{
Fit a rulefit model for survival outcomes
}
