% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_reg.R
\name{SurvModel_SurvReg}
\alias{SurvModel_SurvReg}
\title{SurvModel_SurvReg}
\usage{
SurvModel_SurvReg(data, expvars, timevar, eventvar, dist = "exponential")
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data (needs to be 0/1)}

\item{dist}{distribution for the parametric model, one of "weibull", "exponential", "gaussian", "logistic","lognormal" or "loglogistic".}
}
\value{
a list containing the following objects:
survregOut: fitted survreg model object after forward selection,
predicttrain: predicted survival probabilities matrix for training data (rows=times, cols=observations),
times.interest: unique times for which the probabilities are calculated (including 0),
expvars: explanatory variables included in the final selected model,
varprof: profile of explanatory variables originally provided.
}
\description{
Fit parametric survival models using survival::survreg with forward selection.
}
