% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{SurvModel_xgboost}
\alias{SurvModel_xgboost}
\title{SurvModel_xgboost}
\usage{
SurvModel_xgboost(data, expvars, timevar, eventvar)
}
\arguments{
\item{data}{data frame with explanatory and outcome variables}

\item{expvars}{character vector of names of explanatory variables in data}

\item{timevar}{character name of time variable in data}

\item{eventvar}{character name of event variable in data}
}
\value{
a list containing the following objects.
expvars: vector of explanatory variables,
learner: fitted model object,
estSURVTrain: predicted probabilities,
datatrainProf: sample data for internal use
}
\description{
Fit a xgboost model for survival outcomes using xgboost
}
