% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvModelsFunctions.R
\name{timedepConcordance}
\alias{timedepConcordance}
\title{timedepConcordance}
\usage{
timedepConcordance(predsurv, predsurvtimes, obstimes, obsevents, ctimes = NULL)
}
\arguments{
\item{predsurv}{Predicted matrix of probabilities}

\item{predsurvtimes}{The times for which the probabilities are predicted}

\item{obstimes}{Observed times}

\item{obsevents}{Observed event indicator}

\item{ctimes}{The times at which the concordance index is calculated}
}
\value{
output from the "pec::cindex' function.
}
\description{
Get time dependent concordance
}
