% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_metrics.R
\name{timedepConcordanceCR}
\alias{timedepConcordanceCR}
\title{timedepConcordanceCR}
\usage{
timedepConcordanceCR(
  SurvObj,
  Predictions,
  time,
  cause = 1,
  TestMat = NULL,
  pred_times = NULL
)
}
\arguments{
\item{SurvObj}{survival object (Surv(time, status))}

\item{Predictions}{matrix of predicted CIFs (rows=observations, cols=times) for the cause of interest}

\item{time}{evaluation time point}

\item{cause}{cause of interest (1 or 2)}

\item{TestMat}{test dataset (optional, used if formula needs predictors)}

\item{pred_times}{optional numeric vector of times corresponding to columns in `Predictions`}
}
\value{
concordance index value
}
\description{
Get time dependent concordance for competing risk outcomes (only event 1)
}
