% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRModelsFunctions.R
\name{timedepConcordanceCR}
\alias{timedepConcordanceCR}
\title{timedepConcordanceCR}
\usage{
timedepConcordanceCR(predCIF, predCIFtimes, obstimes, obsevents, TestMat)
}
\arguments{
\item{predCIF}{Predicted matrix of probabilities (CIFs)}

\item{predCIFtimes}{The times for which the probabilities are predicted}

\item{obstimes}{Observed times}

\item{obsevents}{Observed event indicator}

\item{TestMat}{test dataset}
}
\value{
output from the "pec::cindex' function.
}
\description{
Get time dependent concordance for competing risk outcomes (only evemt1)
}
