% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_metrics.R
\name{timedepConcordanceCR}
\alias{timedepConcordanceCR}
\title{timedepConcordanceCR}
\usage{
timedepConcordanceCR(
  predCIF,
  predCIFtimes,
  obstimes,
  obsevents,
  TestMat = NULL
)
}
\arguments{
\item{predCIF}{Predicted matrix of probabilities (CIFs, rows=observations, cols=times)}

\item{predCIFtimes}{The times for which the probabilities are predicted}

\item{obstimes}{Observed times vector}

\item{obsevents}{Observed event indicator vector (0=censored, 1=event of interest, 2=competing event)}

\item{TestMat}{test dataset (optional, used if formula needs predictors)}
}
\value{
output from the "pec::cindex' function.
}
\description{
Get time dependent concordance for competing risk outcomes (only event 1)
}
