% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{UndoZeroOneScalerApplierData}
\alias{UndoZeroOneScalerApplierData}
\title{UndoZeroOneScalerApplierData}
\usage{
UndoZeroOneScalerApplierData(data, mins, maxs)
}
\description{
Un-scale numeric variables using given mins and maxs
}
\details{
@export
}
