% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_deepsurv.R, R/surv_deepsurv.R
\name{unpack_weights}
\alias{unpack_weights}
\title{Unpack Weights}
\usage{
unpack_weights(w_vec, n_in, n_hidden)

unpack_weights(w_vec, n_in, n_hidden)
}
\arguments{
\item{w_vec}{Weight vector.}

\item{n_in}{Number of input features.}

\item{n_hidden}{Number of hidden units.}
}
\value{
A list of weight matrices and bias vectors.

A list of weight matrices and bias vectors.
}
\description{
Unpacks a weight vector into weight matrices and bias vectors.

Unpacks a weight vector into weight matrices and bias vectors.
}
\keyword{internal}
