% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{VariableProfile}
\alias{VariableProfile}
\title{VariableProfile}
\usage{
VariableProfile(data, expvars)
}
\arguments{
\item{data}{data frame}

\item{expvars}{character vector of variable names to profile}
}
\value{
a list named by expvars, containing tables for factors/characters or min/max for numerics.
}
\description{
Creates a profile of variables in a dataset, summarizing factors and numeric ranges.
}
