% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAndDictionaryFunctions.R
\name{ZeroOneScalerApplierData}
\alias{ZeroOneScalerApplierData}
\title{ZeroOneScalerApplierData}
\usage{
ZeroOneScalerApplierData(data, mins, maxs)
}
\arguments{
\item{data}{data frame with data}

\item{mins}{a vector with the minimum values of each variable}

\item{maxs}{a vector with the maximum values of each variable}
}
\value{
a data frame with the data


 @export
}
\description{
Scale numeric variables to the range of 0 and 1 using given mins and maxs
}
