% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{ZeroOneScalerData}
\alias{ZeroOneScalerData}
\title{ZeroOneScalerData}
\usage{
ZeroOneScalerData(data)
}
\arguments{
\item{data}{Data frame with numeric columns to scale.}
}
\value{
A list containing:
  - data: The data frame with numeric columns scaled.
  - minxvec: A named numeric vector of the minimum values used for scaling each column.
  - maxxvec: A named numeric vector of the maximum values used for scaling each column.
}
\description{
Scale numeric variables in a data frame to the range [0, 1].
Records the min and max values used for scaling.
}
